pyproject.toml
[tool.poetry]
name = "nilva"
version = "0.1.0"
description = ""
authors = ["Sajad <imsajtj@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.1"
drf-spectacular = "^0.27.2"
redis = "^5.0.6"
django = "^5.0.6"
django-redis = "^5.4.0"
docker = "^7.1.0"
psycopg2 = "^2.9.9"
requests = "2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
